requirements.txt
git+https://github.com/dolfies/discord.py-self.git
python-dotenv
google-genai
tzdata
